requirements.txt
# Lab Management System (LMS) - Core Dependencies
# Web Framework
fastapi==0.123.4
uvicorn[standard]==0.20.0
gunicorn==20.1.0

# Database
sqlalchemy==1.4.23
psycopg2-binary==2.9.11

# Data Validation
pydantic==2.12.5

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# File Upload & Forms
python-multipart==0.0.6

# Templates
jinja2==3.1.2

# Configuration
python-dotenv==1.0.0
alembic==1.11.1
weasyprint==57.0
